+++ packages/components/atoms/f-error-message/test-utils/component-objects/f-error-message.component.js
@@ -1,8 +1,15 @@
 const Page = require('@justeat/f-wdio-utils/src/page.object');
+const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 
 module.exports = class ErrorMessage extends Page {
     constructor() {
         super('atom', 'error-message-component');
+    }
+
+    load () {
+        const pageUrl = buildUrl(this.componentType, this.componentName, this.path);
+        this.open(pageUrl);
+        this.waitForComponent();
     }
 
     get component () { return $('[data-test-id="error-message-component"]'); }