+++ packages/components/organisms/f-checkout/test/visual/f-checkout-guest.visual.mobile.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const Checkout = require('../../test-utils/component-objects/f-checkout.component');
 
 let checkout = new Checkout();
@@ -11,11 +10,8 @@ describe('f-checkout - Collection - Guest - Mobile Visual Tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Is ASAP available', true);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
     });
 
     it('should display the component base state.', () => {
@@ -50,9 +46,7 @@ describe('f-checkout - Collection - Guest - Mobile Visual Tests', () => {
     it('should display the "Duplicate Order Warning" modal', () => {
         // Arrange
         checkout.withQuery('&knob-Place Order Errors', 'duplicate');
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         checkout.setFieldValue('firstName', 'Jerry');
         checkout.setFieldValue('lastName', 'Jazzman');
         checkout.setFieldValue('emailAddress', 'jerry.jazzman@ronniescotts.co.uk');
@@ -74,11 +68,8 @@ describe('f-checkout - Collection - Guest - isAsapAvailable: false Mobile Visual
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Is ASAP available', false);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
     });
 
     it('should display the pre-order warning.', () => {
@@ -96,11 +87,8 @@ describe('f-checkout - Delivery - Guest - Mobile Visual Tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Is ASAP available', true);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
     });
 
     it('should display the delivery f-checkout component guest base state.', () => {
@@ -136,9 +124,7 @@ describe('f-checkout - Delivery - Guest - Mobile Visual Tests', () => {
     it('should display the "Duplicate Order Warning" modal', () => {
         // Arrange
         checkout.withQuery('&knob-Place Order Errors', 'duplicate');
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         checkout.setFieldValue('firstName', 'Jerry');
         checkout.setFieldValue('lastName', 'Jazzman');
         const addressInfo = {
@@ -166,11 +152,8 @@ describe('f-checkout - Delivery - Guest - isAsapAvailable: false Mobile Visual T
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Is ASAP available', false);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
     });
 
     it('should display the pre-order warning.', () => {
@@ -187,11 +170,8 @@ describe('f-checkout - Dine In - Guest - Mobile Visual Tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Is ASAP available', false);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
     });
 
     it('should display the component base state.', () => {
@@ -232,11 +212,8 @@ describe('f-checkout - Dine In - Guest - Mobile Visual Tests', () => {
             .withQuery('&knob-Is ASAP available', false)
             .withQuery('&knob-Place Order Errors', 'duplicate');
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         checkout.goToPayment();
 
         // Assert