+++ packages/components/organisms/f-cookie-banner/src/components/CookieBanner.vue
@@ -149,7 +149,7 @@ export default {
         const copy = localeConfig.messages;
         const consentCookieName = 'je-cookieConsent';
         const legacyConsentCookieName = 'je-banner_cookie';
-        const reopenLinkToBottom = this.isBodyHeightLessThanWindowHeight;
+        const reopenLinkToBottom = this.isBodyHeightLessThanWindowHeight();
 
         return {
             config: { ...localeConfig },
@@ -170,17 +170,6 @@ export default {
          */
         legacyBanner () {
             return this.shouldShowLegacyBanner === null ? this.config.displayLegacy : this.shouldShowLegacyBanner;
-        },
-
-        /**
-         * Check to see if we need to absolute position reopen link.
-         * * @returns {Boolean}
-         */
-        isBodyHeightLessThanWindowHeight () {
-            if (typeof window === 'object') {
-                return window.innerHeight - document.body.offsetHeight > 0;
-            }
-            return false;
         }
     },
 
@@ -208,6 +197,7 @@ export default {
                 this.setLegacyCookieBannerCookie();
                 this.dataLayerPush('full');
                 this.resendEvents();
+                this.dispatchCustomEvent();
                 this.hideAllBanners();
             });
         },
@@ -222,6 +212,7 @@ export default {
                 this.dataLayerPush('necessary');
                 this.resendEvents();
                 this.removeUnnecessaryCookies();
+                this.dispatchCustomEvent();
                 this.hideAllBanners();
             });
         },
@@ -348,6 +339,31 @@ export default {
                     break;
                 }
             }
+        },
+
+        /**
+         * Check to see if we need to absolute position reopen link.
+         * * @returns {Boolean}
+         */
+        isBodyHeightLessThanWindowHeight () {
+            if (typeof window === 'object') {
+                return window.innerHeight - document.body.offsetHeight > 0;
+            }
+            return false;
+        },
+
+        /**
+         * Dispatch custom window event
+         * */
+        dispatchCustomEvent () {
+            if (typeof window === 'object') {
+                window.dispatchEvent(new CustomEvent('f-cookie-banner-accepted', {
+                    detail: {
+                        consentCookieName: this.consentCookieName,
+                        cookieExclusionList: this.config.cookieExclusionList
+                    }
+                }));
+            }
         }
     }
 };
@@ -381,7 +397,6 @@ export default {
 
     .c-cookieBanner-title {
         @include font-size(heading-m);
-        font-weight: $font-weight-bold;
         margin: spacing() 0;
         padding: 0;
         color: $color-content-default;