+++ packages/components/organisms/f-cookie-banner/test-utils/component-objects/f-cookieConsentBanner.component.js
@@ -1,5 +1,6 @@
 
 const Page = require('@justeat/f-wdio-utils/src/page.object');
+const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 
 module.exports = class CookieBanner extends Page {
     constructor() {
@@ -17,6 +18,14 @@ module.exports = class CookieBanner extends Page {
     get cookieAcceptNecessaryButton () { return this.component.$('[data-test-id="accept-necessary-cookies-button"]'); }
 
     get componentContent () { return $('[data-test-id="cookieBannerContent"]'); }
+
+    load () {
+        const pageUrl = buildUrl(this.componentType, this.componentName, this.path);
+        this.open(pageUrl);
+        browser.deleteAllCookies();
+        browser.refresh();
+        this.waitForComponent();
+    }
 
     open (url) {
         super.open(url);