+++ packages/components/organisms/f-header/test/visual/f-header.visual.mobile.spec.js
@@ -1,7 +1,5 @@
 import forEach from 'mocha-each';
 
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
-
 const Header = require('../../test-utils/component-objects/f-header.component');
 
 let header;
@@ -25,11 +23,9 @@ describe('Shared - f-header component tests', () => {
             if (!isLoggedIn) {
                 header.withQuery('&knob-User info', isLoggedIn);
             }
-            const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
 
             // Act
-            header.open(pageUrl);
-            header.waitForComponent();
+            header.load();
             header.openMobileNavigationBar();
 
             // Assert
@@ -42,11 +38,9 @@ describe('Shared - f-header component tests', () => {
             header = new Header();
             header.withQuery('&knob-Locale', 'en-GB');
             header.withQuery('&knob-Header theme', theme);
-            const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
 
             // Act
-            header.open(pageUrl);
-            header.waitForComponent();
+            header.load();
 
             // Assert
             browser.percyScreenshot(`f-header - Theme colours - ${theme}`, 'mobile');
@@ -55,11 +49,9 @@ describe('Shared - f-header component tests', () => {
     it('should display all avalible countries', () => {
         header = new Header();
         header.withQuery('&knob-Locale', 'en-GB');
-        const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
 
         // Act
-        header.open(pageUrl);
-        header.waitForComponent();
+        header.load();
         header.openMobileNavigationBar();
         header.openCountrySelector();
 
@@ -72,11 +64,9 @@ describe('Shared - f-header component tests', () => {
             header = new Header();
             header.withQuery('&knob-Locale', 'en-GB');
             header.withQuery(`&knob-${knobName}`, 'false');
-            const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
 
             // Act
-            header.open(pageUrl);
-            header.waitForComponent();
+            header.load();
             header.openMobileNavigationBar();
 
             // Assert